Add files via upload
In number theory, two integers a and b are said to be coprime if the only positive integer that divides both of them is 1.
Write a function coprime_test_loop() that asks the user for two numbers. This function will pass those two numbers onto a second function coprime(a,b) which will return true or false depending on whether or not the numbers are coprime. The function coprime_test_loop() will print out a message indicating the result. It will then ask the user for another pair of numbers and query coprime(a,b) again. It will continue this loop until the user indicates that they wish to exit the program.